Support half precision for VQ and FSQ

+++ vector_quantize_pytorch/vector_quantize_pytorch.py
@@ -478,7 +478,7 @@ class EuclideanCodebook(Module):
         needs_codebook_dim = x.ndim < 4
         sample_codebook_temp = default(sample_codebook_temp, self.sample_codebook_temp)
 
-        x = x.float()
+        # x = x.float() removed for other precision
 
         if needs_codebook_dim:
             x = rearrange(x, '... -> 1 ...')
@@ -668,7 +668,7 @@ class CosineSimCodebook(Module):
         needs_codebook_dim = x.ndim < 4
         sample_codebook_temp = default(sample_codebook_temp, self.sample_codebook_temp)
 
-        x = x.float()
+        # x = x.float() removed for other precision
 
         if needs_codebook_dim:
             x = rearrange(x, '... -> 1 ...')
